async-asr-newWebSocket/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
numpy==1.18.4
praat-textgrids==1.3.1
PyAudio==0.2.11
PyQt5==5.14.2
PyQt5-sip==12.7.2
pysrt==1.1.2
python-slugify==4.0.0
python-vlc==3.0.9113
requests==2.23.0
requests-futures==1.0.0
text-unidecode==1.3
urllib3==1.25.9
librosa==0.8.0
pydub==0.24.1
pysubs2==0.2.4
pyqtgraph==0.11.0
spacy==2.3.2
webcolors==1.3
sox==1.4.2
PySimpleGUI==4.49.0
spacy==2.3.2
